add planning factor panel
