+++ DLRLayout.json
@@ -3787,7 +3787,8 @@
       "imageMode": {},
       "followTf": "base_link"
     },
-    "AutowareFoxgloveConverter.DLR Diagnostics Result!1rdd8zb": {},
+    "AutowareFoxgloveConverter.DLR Diagnostics!3l4lita": {},
+    "AutowareFoxgloveConverter.DLR Planning Factor!4kmxj8t": {},
     "AutowareFoxgloveConverter.Vehicle Info!3d09irc": {}
   },
   "globalVariables": {
@@ -3795,12 +3796,17 @@
   },
   "userNodes": {},
   "playbackConfig": {
-    "speed": 1
+    "speed": 5
   },
   "layout": {
     "first": "3D!46c57bh",
     "second": {
-      "first": "AutowareFoxgloveConverter.DLR Diagnostics Result!1rdd8zb",
+      "first": {
+        "first": "AutowareFoxgloveConverter.DLR Diagnostics!3l4lita",
+        "second": "AutowareFoxgloveConverter.DLR Planning Factor!4kmxj8t",
+        "direction": "column",
+        "splitPercentage": 59.53827460510328
+      },
       "second": "AutowareFoxgloveConverter.Vehicle Info!3d09irc",
       "direction": "column",
       "splitPercentage": 81.76850471932438
